requirements.txt
alembic==1.10.2
aniso8601==9.0.1
bcrypt==4.0.1
certifi==2022.9.24
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
Pillow==9.4.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.6
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wincertstore==0.2
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
